Revert "Updated to use loop without argument and UInt64 for attempt number/limit"
This reverts commit 3b2ae097a3c5d7280b27575d66bbb076425c5ebf.

+++ spec/retriable_spec.cr
@@ -26,7 +26,7 @@ describe Retriable do
       tries = 0
       time_table = [] of Time::Span
 
-      handler = ->(ex : Exception, attempt : UInt64, elapsed_time : Time::Span, next_interval : Time::Span) do
+      handler = ->(ex : Exception, attempt : Int32, elapsed_time : Time::Span, next_interval : Time::Span) do
         time_table << next_interval
       end
 
@@ -199,9 +199,9 @@ describe Retriable do
       total_tries = 6
 
       tries = 0
-      time_table = {} of UInt64 => Time::Span
+      time_table = {} of Int32 => Time::Span
 
-      handler = ->(ex : Exception, attempt : UInt64, elapsed_time : Time::Span, next_interval : Time::Span) do
+      handler = ->(ex : Exception, attempt : Int32, elapsed_time : Time::Span, next_interval : Time::Span) do
         ex.should be_a ArgumentError
         time_table[attempt] = next_interval
       end
@@ -234,9 +234,9 @@ describe Retriable do
 
     it "has a max interval of 1.5 seconds" do
       tries = 0
-      time_table = {} of UInt64 => Time::Span
+      time_table = {} of Int32 => Time::Span
 
-      handler = ->(ex : Exception, attempt : UInt64, elapsed_time : Time::Span, next_interval : Time::Span) do
+      handler = ->(ex : Exception, attempt : Int32, elapsed_time : Time::Span, next_interval : Time::Span) do
         time_table[attempt] = next_interval
       end
 
@@ -272,9 +272,9 @@ describe Retriable do
       }
 
       tries = 0
-      time_table = {} of UInt64 => Time::Span
+      time_table = {} of Int32 => Time::Span
 
-      handler = ->(ex : Exception, attempt : UInt64, elapsed_time : Time::Span, next_interval : Time::Span) do
+      handler = ->(ex : Exception, attempt : Int32, elapsed_time : Time::Span, next_interval : Time::Span) do
         time_table[attempt] = next_interval
       end
 
@@ -311,7 +311,7 @@ describe Retriable do
     end
 
     it "works with a hash exception list where the value is a Proc matcher" do
-      ex_matches = ->(ex : Exception, attempt : UInt64, elapsed_time : Time::Span, next_interval : Time::Span) do
+      ex_matches = ->(ex : Exception, attempt : Int32, elapsed_time : Time::Span, next_interval : Time::Span) do
         ex.should be_a TestError
         attempt.should be <= 3
         ex.message == "something went wrong"
@@ -328,7 +328,7 @@ describe Retriable do
     end
 
     it "works with a Proc matcher" do
-      ex_matches = ->(ex : Exception, attempt : UInt64, elapsed_time : Time::Span, next_interval : Time::Span) do
+      ex_matches = ->(ex : Exception, attempt : Int32, elapsed_time : Time::Span, next_interval : Time::Span) do
         ex.should be_a TestError
         attempt.should be <= 3
         ex.class <= TestError && ex.message == "something went wrong"
@@ -378,9 +378,9 @@ describe Retriable do
 
     it "works with a hash exception list where the values are exception message patterns" do
       tries = 0
-      exceptions = {} of UInt64 => Exception
+      exceptions = {} of Int32 => Exception
 
-      handler = ->(ex : Exception, attempt : UInt64, elapsed_time : Time::Span, next_interval : Time::Span) do
+      handler = ->(ex : Exception, attempt : Int32, elapsed_time : Time::Span, next_interval : Time::Span) do
         exceptions[attempt] = ex
       end
 
@@ -424,9 +424,9 @@ describe Retriable do
     )
 
     tries = 0
-    time_table = {} of UInt64 => Time::Span
+    time_table = {} of Int32 => Time::Span
 
-    handler = ->(ex : Exception, attempt : UInt64, elapsed_time : Time::Span, next_interval : Time::Span) do
+    handler = ->(ex : Exception, attempt : Int32, elapsed_time : Time::Span, next_interval : Time::Span) do
       time_table[attempt] = next_interval
     end
